[Serve] Add `policy` config to `AutoscalingConfig`
