+++ python/ray/serve/config.py
@@ -1,8 +1,11 @@
+import binascii
 import logging
 import warnings
+from base64 import b64decode, b64encode
 from enum import Enum
 from typing import Any, Callable, List, Optional, Union
 
+from ray import cloudpickle
 from ray._private.pydantic_compat import (
     BaseModel,
     NonNegativeFloat,
@@ -22,6 +25,7 @@ from ray.serve._private.constants import (
 from ray.util.annotations import Deprecated, PublicAPI
 
 logger = logging.getLogger(SERVE_LOGGER_NAME)
+DEFAULT_AUTOSCALING_POLICY = "ray.serve.autoscaling_policy:BasicAutoscalingPolicy"
 
 
 @PublicAPI(stability="stable")
@@ -60,6 +64,9 @@ class AutoscalingConfig(BaseModel):
     # How long to wait before scaling up replicas
     upscale_delay_s: NonNegativeFloat = 30.0
 
+    # Custom autoscaling config. Defaulting to the request based autoscaler.
+    policy: Union[str, Callable] = DEFAULT_AUTOSCALING_POLICY
+
     @validator("max_replicas", always=True)
     def replicas_settings_valid(cls, max_replicas, values):
         min_replicas = values.get("min_replicas")
@@ -83,6 +90,33 @@ class AutoscalingConfig(BaseModel):
                 )
 
         return max_replicas
+
+    @validator("policy", always=True)
+    def serialize_policy(cls, policy, values):
+        """Serialize policy to be cloudpickled and base64 encoded utf8 string.
+
+        If policy is a string, it is possible to be already serialized when constructed
+        from protobuf. In this case, we just return the already serialized string. Else
+        we import the policy and serialize it before return.
+        """
+        if isinstance(policy, Callable):
+            imported_policy = policy
+        elif isinstance(policy, str):
+            try:
+                assert b64encode(b64decode(policy)).decode("utf-8") == policy
+                return policy
+            except binascii.Error:
+                imported_policy = import_attr(policy)
+            except AssertionError:
+                imported_policy = import_attr(policy)
+        else:
+            return policy
+
+        return b64encode(cloudpickle.dumps(imported_policy)).decode("utf-8")
+
+    def get_policy(self) -> Callable:
+        """Deserialize policy from cloudpicked base64 encoded utf8 string."""
+        return cloudpickle.loads(b64decode(self.policy))
 
     def get_upscale_smoothing_factor(self) -> PositiveFloat:
         return self.upscale_smoothing_factor or self.smoothing_factor
